requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0
prophet==1.1.4
cmdstanpy==1.2.0
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
eventlet==0.33.3
gunicorn==21.2.0
"google-generativeai==0.8.3"
"requests>=2.25.0"
"protobuf>=3.20.0"